Test commit with Atom Editor
Tried making changes in Atom and learning how to commit them. Can’t
commit in Atom (I don’t think, will investigate.), so committing in
GitHub desktop. I did not create a branch to work on in Atom, so I
believe this is going to commit directly to the master.